requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langgraph>=0.0.20
python-dotenv>=1.0.0
pypdf>=3.15.1
ebooklib>=0.18.0
beautifulsoup4>=4.12.2
markdown>=3.5
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
chromadb>=0.4.18
pdfplumber>=0.10.2
